refactor: 消除重复函数，抽取共享模块 _common.py
- 创建 _common.py 集中存放 load_cookies/load_config/get_current_os/parse_output_path
- main.py 删除 6 个重复函数，改为从 _common 和 favorite_ops 导入
- fetch_collections.py 删除 4 个重复函数，改为从 _common 导入
- get_collections.py 标记 deprecated，改为从 fetch_collections 导入
- debug_page.py 删除 load_cookies，改为从 _common 导入
- favorite_ops.py _load_cookies 改为从 _common 导入
