+++ src/zhihu_collections/main.py
@@ -9,116 +9,16 @@ import re
 from tqdm import tqdm
 from datetime import datetime
 from zhihu_collections.utils import filter_title_str
+from zhihu_collections._common import load_config, load_cookies, get_current_os, parse_output_path
+from zhihu_collections.favorite_ops import parse_article_info, unfavorite_content, remove_from_collection
 import json
 import logging
 import traceback
-import platform
-import pathlib
 
 from markdownify import MarkdownConverter
 
 
 # 读取配置文件
-def load_config():
-    try:
-        with open("config.json", "r", encoding="utf-8") as f:
-            config = json.load(f)
-            return config
-    except FileNotFoundError:
-        print("未找到config.json文件，尝试读取旧版zhihuUrls.json文件")
-        try:
-            with open("zhihuUrls.json", "r", encoding="utf-8") as f:
-                urls = json.load(f)
-                return {"zhihuUrls": urls, "outputPath": "", "os": ""}
-        except FileNotFoundError:
-            print("未找到配置文件，请创建config.json文件并配置收藏夹信息")
-            return {"zhihuUrls": [], "outputPath": "", "os": ""}
-
-
-# 获取当前操作系统类型
-def get_current_os():
-    system = platform.system().lower()
-    if system == "windows":
-        return "windows"
-    elif system == "darwin":
-        return "macos"
-    elif system == "linux":
-        return "linux"
-    else:
-        return "unknown"
-
-
-# 解析路径，根据操作系统类型处理
-def parse_output_path(path_str, os_type):
-    if not path_str:
-        return None
-
-    # 如果没有指定os，则自动检测
-    if not os_type:
-        os_type = get_current_os()
-
-    try:
-        if os_type.lower() == "windows":
-            # Windows路径处理
-            # 支持 D:\path\to\folder 或 D:/path/to/folder 格式
-            path_str = path_str.replace("/", "\\")
-            return pathlib.Path(path_str).resolve()
-        elif os_type.lower() in [
-            "linux",
-            "freebsd",
-            "openbsd",
-            "netbsd",
-            "solaris",
-            "aix",
-        ]:
-            # Unix-like系统路径处理
-            # 支持 /usr/local/lib 格式
-            if path_str.startswith("~"):
-                path_str = os.path.expanduser(path_str)
-            return pathlib.Path(path_str).resolve()
-        elif os_type.lower() in ["macos", "darwin"]:
-            # macOS路径处理
-            # 支持 /Users/username/Documents 或 ~/Documents 格式
-            if path_str.startswith("~"):
-                path_str = os.path.expanduser(path_str)
-            return pathlib.Path(path_str).resolve()
-        elif os_type.lower() in ["cygwin", "msys"]:
-            # Cygwin/MSYS环境路径处理
-            # 支持 /cygdrive/c/path 或 /c/path 格式
-            if path_str.startswith("/cygdrive/"):
-                # Cygwin格式: /cygdrive/c/path -> C:\path
-                drive_path = path_str[10:]  # 移除 /cygdrive/
-                if len(drive_path) >= 2 and drive_path[1] == "/":
-                    path_str = (
-                        drive_path[0].upper() + ":" + drive_path[1:].replace("/", "\\")
-                    )
-            elif path_str.startswith("/") and len(path_str) >= 3 and path_str[2] == "/":
-                # MSYS格式: /c/path -> C:\path
-                path_str = path_str[1].upper() + ":" + path_str[2:].replace("/", "\\")
-            return pathlib.Path(path_str).resolve()
-        else:
-            # 其他系统，尝试通用处理
-            logging.warning(f"未知操作系统类型: {os_type}，尝试通用路径处理")
-            if path_str.startswith("~"):
-                path_str = os.path.expanduser(path_str)
-            return pathlib.Path(path_str).resolve()
-    except Exception as e:
-        logging.error(f"路径解析失败: {path_str}, 错误: {str(e)}")
-        return None
-
-
-# 读取cookies
-def load_cookies():
-    try:
-        with open("cookies.json", "r", encoding="utf-8") as f:
-            cookies_list = json.load(f)
-        cookies_dict = {}
-        for cookie in cookies_list:
-            cookies_dict[cookie["name"]] = cookie["value"]
-        return cookies_dict
-    except FileNotFoundError:
-        print("未找到cookies.json文件，将使用无登录模式访问（部分内容可能无法获取）")
-        return {}
 
 
 # 全局变量存储当前处理的收藏夹名称
@@ -1225,165 +1125,6 @@ def process_single_collection(collection_name, collection_url):
     # 将收藏夹日志添加到全局日志
     processing_log.append(collection_log)
     print(f"收藏夹 '{collection_name}' 下载完毕")
-
-
-# ──────────────────────────────────────────────
-# 取消收藏功能
-# ──────────────────────────────────────────────
-
-
-def parse_article_info(url):
-    """
-    解析文章URL，提取 content_id 和 content_type
-
-    支持格式：
-      - https://www.zhihu.com/question/xxx/answer/yyy   → answer, yyy
-      - https://zhuanlan.zhihu.com/p/xxx                  → article, xxx
-      - https://www.zhihu.com/zvideo/xxx                  → zvideo, xxx
-      - https://www.zhihu.com/pin/xxx                     → pin, xxx
-    """
-    import re
-
-    # 匹配回答: /question/xxx/answer/yyy 或 /question/xxx/answer/yyy?...
-    m = re.search(r"/question/\d+/answer/(\d+)", url)
-    if m:
-        return m.group(1), "answer"
-
-    # 匹配专栏文章: /p/xxx
-    m = re.search(r"zhuanlan\.zhihu\.com/p/(\d+)", url)
-    if m:
-        return m.group(1), "article"
-
-    # 匹配视频: /zvideo/xxx
-    m = re.search(r"/zvideo/(\d+)", url)
-    if m:
-        return m.group(1), "zvideo"
-
-    # 匹配想法: /pin/xxx
-    m = re.search(r"/pin/(\d+)", url)
-    if m:
-        return m.group(1), "pin"
-
-    # 兜底：取URL最后一段数字
-    m = re.search(r"/(\d+)(?:\?|#|$)", url)
-    if m:
-        return m.group(1), "unknown"
-
-    return None, None
-
-
-def unfavorite_content(content_type, content_id):
-    """
-    取消收藏指定内容（从收藏夹中移除）
-
-    通过 Playwright 浏览器抓包确定的正确 API 端点：
-      DELETE https://www.zhihu.com/api/v4/collections/contents/{content_type}/{content_id}
-
-    验证结果：
-      - DELETE → 200 {"data":[...], "success":true}  ✅ 取消收藏
-      - POST   → 200 {"collection":{...}, "success":true}  ✅ 添加收藏
-
-    关键请求头：
-      - X-Xsrftoken: CSRF token（从 cookies 中的 _xsrf 获取）
-      - x-zse-93: "101_3_3.0"
-
-    :param content_type: 内容类型 (answer/article/zvideo/pin)
-    :param content_id: 内容ID
-    :return: (success: bool, message: str)
-    """
-    api_url = (
-        f"https://www.zhihu.com/api/v4/collections/contents/{content_type}/{content_id}"
-    )
-
-    # 获取 CSRF token
-    xsrf_token = cookies.get("_xsrf", "") if isinstance(cookies, dict) else ""
-
-    # 专用请求头（从浏览器拦截复制）
-    unfav_headers = {
-        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36",
-        "Accept": "*/*",
-        "Accept-Language": "zh-CN,zh;q=0.9",
-        "Referer": "https://www.zhihu.com/",
-        "x-requested-with": "fetch",
-        "x-zse-93": "101_3_3.0",
-        "Content-Type": "application/json",
-    }
-    if xsrf_token:
-        unfav_headers["X-Xsrftoken"] = xsrf_token
-
-    logging.info(f"取消收藏: DELETE {api_url}")
-
-    try:
-        resp = requests.delete(
-            api_url,
-            headers=unfav_headers,
-            cookies=cookies,
-            timeout=30,
-        )
-
-        if resp.status_code == 200:
-            try:
-                result = resp.json()
-                if result.get("success"):
-                    favlists_count = result.get("favlists_count", "?")
-                    logging.info(f"取消收藏成功, 当前收藏数: {favlists_count}")
-                    return True, f"✅ 已取消收藏 (content_id={content_id})"
-                else:
-                    return True, f"✅ 已取消收藏 (content_id={content_id})"
-            except json.JSONDecodeError:
-                return True, f"✅ 已取消收藏 (content_id={content_id})"
-
-        if resp.status_code in (201, 204):
-            return True, f"✅ 已取消收藏 (content_id={content_id})"
-
-        # 如果是403且错误是"未收藏"，那说明本来就没收藏，也算成功
-        if resp.status_code == 403:
-            try:
-                err = resp.json()
-                if "未收藏" in str(err.get("error", {})):
-                    return True, f"✅ 本来就没有收藏该内容 (content_id={content_id})"
-            except:
-                pass
-
-        error_body = resp.text[:300] if resp.text else "无响应体"
-        return False, (
-            f"❌ 取消收藏失败 (HTTP {resp.status_code})\n"
-            f"   响应: {error_body}\n"
-            f"   提示: 可能是Cookies已过期，请重新登录获取 cookies.json"
-        )
-
-    except requests.exceptions.Timeout:
-        return False, f"❌ 请求超时（30s），请检查网络连接"
-    except requests.exceptions.ConnectionError as e:
-        return False, f"❌ 网络连接失败: {str(e)}"
-    except Exception as e:
-        logging.error(f"取消收藏异常: {str(e)}")
-        return False, f"❌ 操作异常: {str(e)}"
-
-
-def remove_from_collection(collection_url, article_url):
-    """
-    从知乎收藏夹中移除一篇文章（取消收藏）
-
-    通过浏览器抓包验证的正确 API：
-      DELETE https://www.zhihu.com/api/v4/collections/contents/{content_type}/{content_id}
-      返回: {"data":[...], "favlists_count":N, "success":true}
-
-    :param collection_url: 收藏夹URL（仅用于日志）
-    :param article_url: 要移除的文章URL
-    :return: (success: bool, message: str)
-    """
-    collection_id = collection_url.split("?")[0].split("/")[-1]
-    content_id, content_type = parse_article_info(article_url)
-
-    if not content_id:
-        return False, f"❌ 无法从URL中解析文章ID: {article_url}"
-
-    logging.info(
-        f"取消收藏: 收藏夹={collection_id}, content_id={content_id}, type={content_type}"
-    )
-
-    return unfavorite_content(content_type, content_id)
 
 
 def remove_articles_from_collection(collection_url, article_urls):
